requirements.txt
openai
PyYaml
marker-pdf
python-dotenv
fastapi
uvicorn
python-multipart